transaction implemented

+++ src/main/java/com/example/librarymanagementsystem/models/Transaction.java
@@ -33,7 +33,7 @@ public class Transaction {
     @UpdateTimestamp
     private Date updatedOn;
 
-    private Double fine;
+    private Integer fine;
 
     @JoinColumn
     @ManyToOne
